user section redesign
